(fga-eps-mds/2023-1-CAPJu-Doc#23) - Padroniza nome das modais da aplicação
